% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pathomove_output.R
\name{check_pathomove_output}
\alias{check_pathomove_output}
\title{Checks the validity of the \code{pathomove_output} class.}
\usage{
check_pathomove_output(object)
}
\arguments{
\item{object}{Instance of the \code{pathomove_output} class.}
}
\value{
Boolean or errors
}
\description{
Checks the validity of the \code{pathomove_output} class.
}
