% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check_cluster.R
\name{check_prepare_cluster}
\alias{check_prepare_cluster}
\title{Prepare the cluster for snevo.}
\usage{
check_prepare_cluster(ssh_con = "some_server", password = "your_password")
}
\arguments{
\item{ssh_con}{Cluster connection.}

\item{password}{Password.}
}
\value{
None, should check and prepare cluster.
}
\description{
Prepare the cluster for snevo.
}
