% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_eco_sim}
\alias{do_eco_sim}
\title{Get movement in ecological time.}
\usage{
do_eco_sim(
  popsize,
  landsize,
  nFood,
  nClusters,
  clusterDispersal,
  maxAct,
  activityRatio,
  pInactive,
  collective,
  sensoryRange,
  tmax,
  scenes
)
}
\arguments{
\item{popsize}{The population size.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{nFood}{The number of food items.}

\item{nClusters}{Number of clusters around which food is generated.}

\item{clusterDispersal}{How dispersed food is around the cluster centre.}

\item{maxAct}{The maximum acitivity.}

\item{activityRatio}{The ratio of activity levels of inactive individuals relative to active individuals.}

\item{pInactive}{The proportion of inactive individuals in the population.}

\item{collective}{Whether to move collectively.}

\item{sensoryRange}{The sensory range of agents.}

\item{tmax}{The integer number of timesteps per scene.}

\item{scenes}{Number of scenes.}
}
\value{
A list with data frames of the population movement.
}
\description{
Run the movement component on a starter population.
}
