% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_simulation}
\alias{do_simulation}
\title{Runs the sociality model simulation.}
\usage{
do_simulation(genmax, tmax, foodClusters, clusterDispersal)
}
\arguments{
\item{genmax}{The maximum number of generations per simulation.}

\item{tmax}{The number of timesteps per generation.}

\item{foodClusters}{Number of clusters around which food is generated.}

\item{clusterDispersal}{How dispersed food is around the cluster centre.}
}
\value{
Nothing. Runs simulation.
}
\description{
Run the simulation using parameters passed as
arguments to the corresponding R function.
}
