% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_social_info.R
\name{get_agent_avoidance}
\alias{get_agent_avoidance}
\title{Get agent avoidance.}
\usage{
get_agent_avoidance(df)
}
\arguments{
\item{df}{A dataframe with scaled decision weights.}
}
\value{
A column of agent avoidance value, where agent avoidance is
the sum of negative scaled agent weights.
}
\description{
Get agent avoidance.
}
