% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_process_networks.R
\name{get_chain_size}
\alias{get_chain_size}
\title{Get transmission chain sizes}
\usage{
get_chain_size(graph)
}
\arguments{
\item{graph}{An object inheriting from class \code{igraph}, typically the output
of \code{get_transmission_chains}.}
}
\value{
A vector of transmission chain sizes.
}
\description{
Get transmission chain sizes
}
