% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_social_info.R
\name{get_functional_variation}
\alias{get_functional_variation}
\title{Get functional variation in movement weights.}
\usage{
get_functional_variation(df)
}
\arguments{
\item{df}{A dataframe with generation, id, and movement weights.}
}
\value{
Nothing. Transforms weights by reference. See data.table.
}
\description{
Get functional variation in movement weights.
}
