% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_handle_data.R
\name{get_move_data}
\alias{get_move_data}
\title{Get sample movement data from \emph{pathomove} output.}
\usage{
get_move_data(object)
}
\arguments{
\item{object}{A \code{pathomove_output} object.}
}
\value{
A \code{data.table} with the positions of all individuals in two
generations: "pre", for the generation immediately before pathogen
introduction (\code{g_patho_init} - 1 in \code{run_pathomove}); and "post", for the
last generation of the simulation (\code{genmax} - 1 in \code{run_pathomove}).
}
\description{
Get sample movement data from \emph{pathomove} output.
}
