% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_process_networks.R
\name{get_networks}
\alias{get_networks}
\title{Get network data from Rds files.}
\usage{
get_networks(output, assoc_threshold = 5)
}
\arguments{
\item{output}{A \code{pathomove_output} object.}

\item{assoc_threshold}{The association count threshold.}
}
\value{
A list of \code{tidygraph} objects.
}
\description{
Get network data from Rds files.
}
