% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_social_info.R
\name{get_si_importance}
\alias{get_si_importance}
\title{Importance of social strategy.}
\usage{
get_si_importance(df)
}
\arguments{
\item{df}{The dataframe with agent id and social weights.}
}
\value{
A dataframe with social strategy importance.
}
\description{
Importance of social strategy.
}
