% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_social_info.R
\name{get_social_strategy}
\alias{get_social_strategy}
\title{Assign a social strategy}
\usage{
get_social_strategy(df)
}
\arguments{
\item{df}{The dataframe with agent id and social weights.}
}
\value{
A dataframe with social strategy assigned.
}
\description{
Assign a social strategy
}
