% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_test_landscape}
\alias{get_test_landscape}
\title{Returns a test landscape.}
\usage{
get_test_landscape(nItems, landsize, nClusters, clusterSpread, regen_time)
}
\arguments{
\item{nItems}{How many items.}

\item{landsize}{Size as a numeric (float).}

\item{nClusters}{How many clusters, an integer value.}

\item{clusterSpread}{Dispersal of items around cluster centres.}

\item{regen_time}{Regeneration time, in timesteps.}
}
\value{
A data frame of the evolved population traits.
}
\description{
Returns a test landscape.
}
