% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMovement}
\alias{getMovement}
\title{Get movement in ecological time.}
\usage{
getMovement(
  popsize,
  landsize,
  nFood,
  nClusters,
  clusterDispersal,
  collective,
  tmax
)
}
\arguments{
\item{popsize}{The population size.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{nFood}{The number of food items.}

\item{nClusters}{Number of clusters around which food is generated.}

\item{clusterDispersal}{How dispersed food is around the cluster centre.}

\item{collective}{Whether to move collectively.}

\item{tmax}{The number of timesteps per generation.}
}
\value{
A list with data frames of the population movement.
}
\description{
Run the movement component on a starter population.
}
