% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_parameters.R
\name{make_parameter_file}
\alias{make_parameter_file}
\title{Write parameters as CSV.}
\usage{
make_parameter_file(replicates, ..., which_file = "this_file.csv")
}
\arguments{
\item{replicates}{How many replicates.}

\item{...}{Pass arguments to \code{run\_noisewalker}.}

\item{which_file}{Where to save these parameters.}
}
\value{
Writes a parameter file.
}
\description{
Write parameters as CSV.
}
