% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pathomove_output.R
\docType{class}
\name{pathomove_output-class}
\alias{pathomove_output-class}
\title{Defines the \code{pathomove_output} class.}
\description{
Defines the \code{pathomove_output} class.
}
\section{Slots}{

\describe{
\item{\code{scenario}}{integer.}

\item{\code{n_gen}}{integer.}

\item{\code{gen_patho_intro}}{integer.}

\item{\code{infections_per_gen}}{integer.}
}}

