% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{plot_pathomove}
\alias{plot_pathomove}
\title{Plot the outcomes of a \emph{pathomove} simulation.}
\usage{
plot_pathomove(x, type = "infections")
}
\arguments{
\item{x}{A \code{pathomove_output} object.}

\item{type}{What type of output to print.
The current option is \code{infections} by default, for a plot of infections in
each recorded generation.}
}
\value{
Shows a \code{ggplot} plot.
}
\description{
Plot the outcomes of a \emph{pathomove} simulation.
}
