% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_handle_data.R
\name{pmv_get_trait_data}
\alias{pmv_get_trait_data}
\title{Get trait data from \emph{pathomove} output.}
\usage{
pmv_get_trait_data(object, scaled_preferences = TRUE)
}
\arguments{
\item{object}{A \code{pathomove_output} object.}

\item{scaled_preferences}{Boolean, whether to scale agents' movement
preferences before returning the trait (i.e., preference) data.}
}
\value{
A \code{data.table} with all reported generations.
}
\description{
Get trait data from \emph{pathomove} output.
}
