% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_pathomove}
\alias{run_pathomove}
\title{Runs the pathomove simulation.}
\usage{
run_pathomove(
  scenario,
  popsize,
  nItems,
  landsize,
  nClusters,
  clusterSpread,
  tmax,
  genmax,
  g_patho_init,
  range_food,
  range_agents,
  range_move,
  handling_time,
  regen_time,
  pTransmit,
  initialInfections,
  costInfect,
  nThreads,
  local_dispersal,
  infect_percent
)
}
\arguments{
\item{scenario}{The pathomove scenario: 1 for ancestral pathogen, 2 for spillover pathogen.}

\item{popsize}{The population size.}

\item{nItems}{How many food items on the landscape.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{nClusters}{Number of clusters around which food is generated.}

\item{clusterSpread}{How dispersed food is around the cluster centre.}

\item{tmax}{The number of timesteps per generation.}

\item{genmax}{The maximum number of generations per simulation.}

\item{g_patho_init}{The generation in which to begin introducing the pathogen.}

\item{range_food}{The sensory range for food.}

\item{range_agents}{The sensory range for agents.}

\item{range_move}{The movement range for agents.}

\item{handling_time}{The handling time.}

\item{regen_time}{The item regeneration time.}

\item{pTransmit}{Probability of transmission.}

\item{costInfect}{The per-timestep cost of pathogen infection.}

\item{nThreads}{How many threads to parallelise over. Set to 1 to run on
the HPC Peregrine cluster.}

\item{local_dispersal}{A boolean value; whether to implement local
(\code{TRUE}) or global (\code{FALSE}) natal dispersal.}

\item{infect_percent}{A boolean value; whether the infection depletes a
percentage of daily energy (\code{TRUE}) or whether a fixed value
(\code{FALSE}) is subtracted from net energy.
For \code{infect_percent = TRUE}, the net energy remaining after \code{T}
timesteps of infection is \code{N * (1 - cost_infect) ^ T}, where \code{N}
is total intake.
For \code{infect_percent = FALSE}, the net energy remaining after \code{T}
timesteps of infection is \code{N - (cost_infect * T)}, where \code{N}
is total intake.}

\item{intialInfections}{Agents infected per event.}
}
\value{
A data frame of the evolved population traits.
}
\description{
Run the simulation using parameters passed as
arguments to the corresponding R function.
}
