% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_pathomove_s4}
\alias{run_pathomove_s4}
\title{Runs the pathomove simulation and return a \code{pathomove_output} object.}
\usage{
run_pathomove_s4(
  scenario,
  popsize,
  nItems,
  landsize,
  nClusters,
  clusterSpread,
  tmax,
  genmax,
  g_patho_init,
  range_food,
  range_agents,
  range_move,
  handling_time,
  regen_time,
  pTransmit,
  initialInfections,
  costInfect,
  nThreads,
  dispersal,
  infect_percent,
  vertical,
  mProb,
  mSize
)
}
\arguments{
\item{scenario}{The pathomove scenario: 0 for no pathogen, 1 for
persistent introduction across generations, and 2 for a single introduction.}

\item{popsize}{The population size.}

\item{nItems}{How many food items on the landscape.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{nClusters}{Number of clusters around which food is generated.}

\item{clusterSpread}{How dispersed food is around the cluster centre.}

\item{tmax}{The number of timesteps per generation.}

\item{genmax}{The maximum number of generations per simulation.}

\item{g_patho_init}{The generation in which to begin introducing the pathogen.}

\item{range_food}{The sensory range for food.}

\item{range_agents}{The sensory range for agents.}

\item{range_move}{The movement range for agents.}

\item{handling_time}{The handling time.}

\item{regen_time}{The item regeneration time.}

\item{pTransmit}{Probability of transmission.}

\item{initialInfections}{Agents infected per event.}

\item{costInfect}{The per-timestep cost of pathogen infection.}

\item{nThreads}{How many threads to parallelise over. Set to 1 to run on
the HPC Peregrine cluster.}

\item{dispersal}{A float value; the standard deviation of a normal
distribution centred on zero, which determines how far away from its parent
each individual is initialised. The standard value is 5 percent of the
landscape size (\code{landsize}), and represents local dispersal.
Setting this to 10 percent is already almost equivalent to global dispersal.}

\item{infect_percent}{A boolean value; whether the infection depletes a
percentage of daily energy (\code{TRUE}) or whether a fixed value
(\code{FALSE}) is subtracted from net energy.
For \code{infect_percent = TRUE}, the net energy remaining after \code{T}
timesteps of infection is \code{N * (1 - cost_infect) ^ T}, where \code{N}
is total intake.
For \code{infect_percent = FALSE}, the net energy remaining after \code{T}
timesteps of infection is \code{N - (cost_infect * T)}, where \code{N}
is total intake.}

\item{vertical}{Should the pathogen be transmitted vertically? Should be
set to \code{TRUE} for a realistic implementation of scenario 3, \emph{single spillover}.}

\item{mProb}{The probability of mutation. The suggested value is 0.01.
While high, this may be more appropriate for a small population; change this
value and \code{popsize} to test the simulation's sensitivity to these values.}

\item{mSize}{Controls the mutational step size, and represents the scale
parameter of a Cauchy distribution.}
}
\value{
An S4 class, \code{pathomove_output}, with simulation outcomes.
}
\description{
Run the simulation using parameters passed as
arguments to the corresponding R function.
}
