% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_summary.R
\docType{methods}
\name{summary.pathomove_output}
\alias{summary.pathomove_output}
\title{Summarise a \code{pathomove_output} object.}
\usage{
\S4method{summary}{pathomove_output}(object)
}
\description{
Summarise a \code{pathomove_output} object.
}
