% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_simulation}
\alias{do_simulation}
\title{Runs the sociality model simulation.}
\usage{
do_simulation(
  popsize,
  genmax,
  tmax,
  nFood,
  foodClusters,
  clusterDispersal,
  landsize,
  competitionCost,
  sensoryRange,
  nScenes,
  stopTime
)
}
\arguments{
\item{popsize}{The population size.}

\item{genmax}{The maximum number of generations per simulation.}

\item{tmax}{The number of timesteps per generation.}

\item{nFood}{The number of food items.}

\item{foodClusters}{Number of clusters around which food is generated.}

\item{clusterDispersal}{How dispersed food is around the cluster centre.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{competitionCost}{Cost of associations.}

\item{sensoryRange}{The sensory range.}

\item{nScenes}{How many scenes.}

\item{stopTime}{The handling time.}
}
\value{
A data frame of the evolved population traits.
}
\description{
Run the simulation using parameters passed as
arguments to the corresponding R function.
}
