% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{export_test_landscapes}
\alias{export_test_landscapes}
\title{Make landscapes with discrete food items in clusters.}
\usage{
export_test_landscapes(
  foodClusters,
  clusterDispersal,
  nFood,
  landsize,
  replicates
)
}
\arguments{
\item{foodClusters}{Number of clusters around which food is generated.}

\item{clusterDispersal}{How dispersed food is around the cluster centre.}

\item{nFood}{The number of food items.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{replicates}{How many replicates.}
}
\value{
Nothing. Runs simulation.
}
\description{
Makes landscape and writes them to file with unique id.
}
