% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_parameters.R
\name{make_parameter_file}
\alias{make_parameter_file}
\title{Write parameters as CSV.}
\usage{
make_parameter_file(
  generations = 1000,
  popsize = 1000,
  tmax = 100,
  clusters = seq(1, 10),
  spread = seq(1, 5),
  replicates = 10,
  which_file = "this_file.csv"
)
}
\arguments{
\item{generations}{Max generations.}

\item{popsize}{Max population size.}

\item{tmax}{Timesteps.}

\item{clusters}{Food clusters.}

\item{spread}{Spread of food clusters.}

\item{replicates}{How many replicates.}

\item{which_file}{Which file to write to.}
}
\value{
Writes a parameter file.
}
\description{
Write parameters as CSV.
}
