% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pathomove_output.R
\docType{class}
\name{pathomove_output-class}
\alias{pathomove_output-class}
\title{Defines the \code{pathomove_output} class.}
\description{
Defines the \code{pathomove_output} class.
}
\section{Slots}{

\describe{
\item{\code{agent_parameters}}{list.}

\item{\code{eco_parameters}}{list.}

\item{\code{generations}}{integer.}

\item{\code{gens_patho_intro}}{integer.}

\item{\code{infections_per_gen}}{integer.}

\item{\code{trait_data}}{list.}

\item{\code{edge_lists}}{list.}

\item{\code{gens_edge_lists}}{list.}
}}

