% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_pathomove}
\alias{run_pathomove}
\title{Runs the pathomove simulation.}
\usage{
run_pathomove(
  scenario,
  popsize,
  nItems,
  landsize,
  nClusters,
  clusterSpread,
  tmax,
  genmax,
  range_food,
  range_agents,
  handling_time,
  regen_time,
  pTransmit,
  initialInfections,
  costInfect,
  nThreads
)
}
\arguments{
\item{scenario}{The pathomove scenario: 1 for ancestral pathogen, 2 for spillover pathogen.}

\item{popsize}{The population size.}

\item{nItems}{How many food items on the landscape.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{nClusters}{Number of clusters around which food is generated.}

\item{clusterSpread}{How dispersed food is around the cluster centre.}

\item{tmax}{The number of timesteps per generation.}

\item{genmax}{The maximum number of generations per simulation.}

\item{range_food}{The sensory range for food.}

\item{range_agents}{The sensory range for agents.}

\item{handling_time}{The handling time.}

\item{regen_time}{The item regeneration time.}

\item{pTransmit}{Probability of transmission.}

\item{costInfect}{Cost infection.}

\item{nThreads}{How many threads to parallelise over.}

\item{intialInfections}{Agents infected per event.}
}
\value{
A data frame of the evolved population traits.
}
\description{
Run the simulation using parameters passed as
arguments to the corresponding R function.
}
