% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_summary.R
\name{summary}
\alias{summary}
\alias{summary.pathomove_output}
\title{Summarise a \code{pathomove_output} object.}
\usage{
summary(x, ...)

\method{summary}{pathomove_output}(x)
}
\arguments{
\item{x}{A \code{pathomove_output} object.}

\item{...}{}
}
\value{
Prints a summary of the \code{pathomove_output} object.
}
\description{
Summarise a \code{pathomove_output} object.
}
